symphonia_common::mpeg

Module audio

Structs

AudioSpecificConfig

MPEG4 Audio Specific Configuration.

Enums

AudioObjectType

MPEG4 audio object types.

Mpeg4AudioChannels

MPEG4 audio channel configuration result.

Mpeg4AudioSampleRate

MPEG4 audio sample rate result.

Functions

get_audio_codec_profile

get_mpeg4_audio_channels_by_config_index

Try to get the set of channels given the configuration index.

get_mpeg4_audio_object_type_by_index

Try to get the audio object type from the given audio object type index.

get_mpeg4_audio_sample_rate_by_index

Try to get the audio sample rate given the sample rate index.